backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
python-multipart==0.0.6
alembic==1.13.1
stripe==8.1.0
requests==2.31.0
python-dotenv==1.0.0
email-validator==2.1.0